requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization and logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
pathlib>=1.0.1